Support NHANES GT3X files
