+++ pyproject.toml
@@ -19,6 +19,7 @@ pytest-resource-path = "^1.3.0"
 mypy = "^0.931"
 flake8 = "^4.0.1"
 pytest-cov = "^3.0.0"
+pydocstyle = "^6.1.1"
 
 [[tool.mypy.overrides]]
 module = ["pandas"]
