+++ tests/conftest.py
@@ -13,6 +13,11 @@ def agdc_file(resource_path_root):
 
 
 @pytest.fixture
+def v1_file(resource_path_root):
+    return resource_path_root / "neo.gt3x"
+
+
+@pytest.fixture
 def calibrated_csv_file(resource_path_root):
     return resource_path_root / "WRIST_rawCalibrated_032Hz.csv"
