+++ tests/test_calibration.py
@@ -1,7 +1,6 @@
 import pytest
 
-from pygt3x.calibrated_reader import CalibratedReader
-from pygt3x.callibration_v2_service import CalibrationV2Service
+from pygt3x.calibration import CalibratedReader, CalibrationV2Service
 from pygt3x.reader import FileReader
 
 test_calibration = {
@@ -38,16 +37,6 @@ def test_read(gt3x_file):
     assert df.mean().X == pytest.approx(-0.3164288263865748)
     assert df.mean().Y == pytest.approx(0.31966339634473695)
     assert df.mean().Z == pytest.approx(0.25388985885877546)
-
-
-def test_read_agdc(agdc_file):
-
-    with FileReader(agdc_file) as reader:
-        calibrated = CalibratedReader(reader)
-        df = calibrated.to_pandas()
-    assert df.mean().X == pytest.approx(1.118586370635872)
-    assert df.mean().Y == pytest.approx(0.4294534356312101)
-    assert df.mean().Z == pytest.approx(-0.08104933668082379)
 
 
 def test_calibrate_32hz(calibrated_dataframe, wrist_dataframe):
